Certificate issuance and renewals work now Also, some refactoring
