+++ worker/cert-monitor.go
@@ -29,6 +29,7 @@ import (
 	"time"
 
 	"github.com/statiko-dev/statiko/appconfig"
+	"github.com/statiko-dev/statiko/certificates"
 	"github.com/statiko-dev/statiko/notifications"
 	"github.com/statiko-dev/statiko/state"
 	"github.com/statiko-dev/statiko/sync"
@@ -58,18 +59,21 @@ func startCertMonitorWorker(ctx context.Context) {
 		// Wait for startup
 		waitForStartup()
 
-		// Run right away
-		err := certMonitorWorker()
-		if err != nil {
-			certMonitorLogger.Println("Worker error:", err)
-		}
-
 		// Run on ticker
 		ticker := time.NewTicker(certMonitorInterval)
 		defer ticker.Stop()
+
+		// Run right away
+		state.Instance.TriggerRefreshCerts()
+
 		for {
 			select {
 			case <-ticker.C:
+				err := certMonitorWorker()
+				if err != nil {
+					certMonitorLogger.Println("Worker error:", err)
+				}
+			case <-state.Instance.RefreshCerts:
 				err := certMonitorWorker()
 				if err != nil {
 					certMonitorLogger.Println("Worker error:", err)
@@ -86,34 +90,38 @@ func startCertMonitorWorker(ctx context.Context) {
 func certMonitorWorker() error {
 	certMonitorLogger.Println("Starting cert-monitor worker")
 
+	// If there's a sync running, wait for it to be done
+	for sync.IsRunning() {
+		certMonitorLogger.Println("Waiting for sync to be complete")
+		time.Sleep(2 * time.Second)
+	}
+
+	// Init variablles
 	now := time.Now()
 	needsSync := false
-
-	// Scan all sites on disk
 	appRoot := appconfig.Config.GetString("appRoot")
 	if !strings.HasSuffix(appRoot, "/") {
 		appRoot += "/"
 	}
-	sites, err := ioutil.ReadDir(appRoot + "sites/")
-	if err != nil {
-		return err
-	}
-	for _, el := range sites {
-		site := el.Name()
 
+	// Go through all sites
+	sites := state.Instance.GetSites()
+
+	// Scan all sites on disk
+	for _, el := range sites {
 		// Check if there's a TLS certificate for this site
-		hasCert, err := utils.FileExists(appRoot + "sites/" + site + "/tls/certificate.pem")
-		if err != nil {
-			return err
-		}
-		if !hasCert {
+		if el.TLS == nil || el.TLS.Type == "" {
 			continue
 		}
 
 		// Read the TLS certificate
-		certMonitorLogger.Println("Reading certificate for site", site)
-		certData, err := ioutil.ReadFile(appRoot + "sites/" + site + "/tls/certificate.pem")
+		certMonitorLogger.Println("Reading certificate for site", el.Domain)
+		certData, err := ioutil.ReadFile(appRoot + "sites/" + el.Domain + "/tls/certificate.pem")
 		if err != nil {
+			if os.IsNotExist(err) {
+				// This should never happen... regardless, just move on to the next site
+				continue
+			}
 			return err
 		}
 		p, _ := pem.Decode(certData)
@@ -125,29 +133,27 @@ func certMonitorWorker() error {
 			return err
 		}
 
-		// Is this certificate self-signed?
-		selfSigned := false
-		if len(cert.Issuer.Organization) > 0 && cert.Issuer.Organization[0] == utils.SelfSignedCertificateIssuer {
-			selfSigned = true
-		}
+		// List of domains
+		domains := append([]string{el.Domain}, el.Aliases...)
 
 		// Check if we sent a notification for expiring certificates already
-		sent, found := certMonitorNotifications[site]
+		sent, found := certMonitorNotifications[el.Domain]
 		if !found {
 			sent = len(certMonitorChecks)
 		}
 
 		// Check expiry date
 		exp := cert.NotAfter
-		if selfSigned {
-			// Certificate is self-signed, so let's just restart the server to have it regenerate if it's got less than 7 days left
-			if exp.Before(now.Add(time.Duration(time.Duration(utils.SelfSignedMinDays*24) * time.Hour))) {
-				certMonitorLogger.Println("Certificate for site", site, "is expiring in less than 7 days; regenerating it")
+		switch el.TLS.Type {
+		case state.TLSCertificateSelfSigned:
+			// Certificate is self-signed, so let's just restart the server to have it regenerate if it's got less than N days left
+			if exp.Before(now.Add(time.Duration(certificates.SelfSignedMinDays*24) * time.Hour)) {
+				certMonitorLogger.Printf("Certificate for site %s is expiring in less than %d days; regenerating it\n", el.Domain, certificates.SelfSignedMinDays)
 
 				// Queue a job
 				job := utils.JobData{
 					Type: utils.JobTypeTLSCertificate,
-					Data: strings.Join(cert.DNSNames, ","),
+					Data: strings.Join(domains, ","),
 				}
 				jobID, err := state.Worker.AddJob(job)
 				if err != nil {
@@ -166,14 +172,15 @@ func certMonitorWorker() error {
 				// We'll queue a sync
 				needsSync = true
 			}
-		} else {
-			// Note: we are assuming 24-hour days, which isn't always correct but it's fine in this case
+		case state.TLSCertificateImported:
+			// Imported certificate
 			for i := 0; i < len(certMonitorChecks); i++ {
 				// If the certificate has expired
+				// Note: we are assuming 24-hour days, which isn't always correct but it's fine in this case
 				if exp.Before(now.Add(time.Duration(certMonitorChecks[i]*24) * time.Hour)) {
 					// If we haven't already sent this notification
 					if i < sent {
-						message := "Certificate for " + site + " "
+						message := "Certificate for " + el.Domain + " "
 						if certMonitorChecks[i] == -2 {
 							message += "has expired over 2 days ago"
 						} else if certMonitorChecks[i] == -1 {
@@ -185,11 +192,44 @@ func certMonitorWorker() error {
 						} else {
 							message += fmt.Sprintf("expires in %d days", certMonitorChecks[i])
 						}
-						certMonitorNotifications[site] = i
+						certMonitorNotifications[el.Domain] = i
 						go notifications.SendNotification(message)
 						break
 					}
 				}
+			}
+		case state.TLSCertificateACME:
+			// Certificate issued by an ACME provider (e.g. Let's Encrypt)
+			// We will need to request a new certificate if it's expiring
+			// Additionally, if the certificate currently on disk was self-signed, it was just temporary, so we need to request the certificate
+			expired := exp.Before(now.Add(time.Duration(certificates.ACMEMinDays*24) * time.Hour))
+			selfSigned := cert.Issuer.Organization != nil &&
+				len(cert.Issuer.Organization) > 0 &&
+				cert.Issuer.Organization[0] == certificates.SelfSignedCertificateIssuer
+			if expired || selfSigned {
+				certMonitorLogger.Printf("Requesting a new certificate for site %s from ACME\n", el.Domain)
+
+				// Queue a job
+				job := utils.JobData{
+					Type: utils.JobTypeACME,
+					Data: strings.Join(domains, ","),
+				}
+				jobID, err := state.Worker.AddJob(job)
+				if err != nil {
+					return err
+				}
+
+				// Wait for the job
+				ch := make(chan error, 1)
+				go state.Worker.WaitForJob(jobID, ch)
+				err = <-ch
+				close(ch)
+				if err != nil {
+					return err
+				}
+
+				// We'll queue a sync
+				needsSync = true
 			}
 		}
 	}
